Dialect tests is only testing postgres, so moving those tests to a postgres directory, and splitting them up across rational filenames
