+++ test/postgres/update-tests.js
@@ -1,0 +1,20 @@
+var Harness = require('./support');
+var post = Harness.definePostTable();
+
+Harness.test({
+  query : post.update({content: 'test'}),
+  pg    : 'UPDATE "post" SET "content" = $1',
+  params: ['test']
+});
+
+Harness.test({
+  query : post.update({content: 'test', userId: 3}),
+  pg    : 'UPDATE "post" SET "content" = $1, "userId" = $2',
+  params: ['test', 3]
+});
+
+Harness.test({
+  query : post.update({content: 'test', userId: 3}).where(post.content.equals('no')),
+  pg    : 'UPDATE "post" SET "content" = $1, "userId" = $2 WHERE ("post"."content" = $3)',
+  params: ['test', 3, 'no']
+});
